# Benchmark for ST_Contains