… presto-geospatial-toolkit/src/main/java/com/facebook/presto/geospatial/GeometryUtils.java