It turns out that 60% of execution time for ST_Contains is spent deserializing geometries from slices. This PR adds a benchmark to show that and optimizes ST_Contains to avoid costly deserialization when possible.

The benchmark runs ST_Contains for a large polygon and 3 points:
* inside the polygon, 
* outside of the polygon, but inside the polygon's bounding box,
* outside of the polygon's bounding box. 

The benchmark runs ST_Contains using both slices and geometries. Here are the results:

```
Benchmark                                                            Mode  Cnt       Score       Error  Units
BenchmarkSTContains.deserialize                                      avgt   20  131919.856 ±  4921.112  ns/op
BenchmarkSTContains.deserializeSimpleGeometry                        avgt   20     486.274 ±    13.489  ns/op
BenchmarkSTContains.stContainsInnerPoint                             avgt   20  227284.106 ± 18774.481  ns/op
BenchmarkSTContains.stContainsInnerPointDeserialized                 avgt   20   66938.920 ±  1570.299  ns/op
BenchmarkSTContains.stContainsInnerPointSimpleGeometry               avgt   20     813.359 ±    34.580  ns/op
BenchmarkSTContains.stContainsOuterPointInEnvelope                   avgt   20  218173.265 ±  9345.754  ns/op
BenchmarkSTContains.stContainsOuterPointInEnvelopeDeserialized       avgt   20   69004.912 ±  2625.112  ns/op
BenchmarkSTContains.stContainsOuterPointInEnvelopeSimpleGeometry     avgt   20     837.565 ±    58.022  ns/op
BenchmarkSTContains.stContainsOuterPointNotInEnvelope                avgt   20  134204.161 ±  5208.045  ns/op
BenchmarkSTContains.stContainsOuterPointNotInEnvelopeDeserialized    avgt   20      30.272 ±     1.232  ns/op
BenchmarkSTContains.stContainsOuterPointNotInEnvelopeSimpleGeometry  avgt   20     643.869 ±    15.003  ns/op
```

For points within the polygon's bounding box, ST_Contains on geometries is almost 3 times faster than on slices. For points outside of the polygon's bounding, the difference is even larger.

ST_Contains runs in two phases: (1) check if bounding box of the left geometry contains bounding box of the right geometry; if not, return false; (2) check if the left geometry contains the right geometry. This is what makes ST_Contains for points outside of polygon's bounding box run faster.

Phase 1 doesn't require full geometries, only their bounding boxes. Hence, it is sufficient to deserialize just that part (read 4 doubles). Making this change greatly speeds up ST_Contains for cases when computation completes in phase 1 and doesn't proceed to phase 2 (e.g. point is outside polygon's bounding box). Benchmark results for the optimized version of the
ST_Contains show 6x improvement for simple geometry and 1000x improvement for
complex geometry.

```
Benchmark                                                            Mode  Cnt       Score       Error  Units
BenchmarkSTContains.deserialize                                      avgt   20  138402.816 ± 11643.730  ns/op
BenchmarkSTContains.deserializeEnvelope                              avgt   20      59.143 ±     2.541  ns/op
BenchmarkSTContains.deserializeEnvelopeSimpleGeometry                avgt   20      61.340 ±     3.933  ns/op
BenchmarkSTContains.deserializeSimpleGeometry                        avgt   20     499.509 ±    22.847  ns/op
BenchmarkSTContains.stContainsInnerPoint                             avgt   20  231484.522 ± 17018.650  ns/op
BenchmarkSTContains.stContainsInnerPointDeserialized                 avgt   20   67700.512 ±  2490.978  ns/op
BenchmarkSTContains.stContainsInnerPointSimpleGeometry               avgt   20     920.532 ±    21.465  ns/op
BenchmarkSTContains.stContainsOuterPointInEnvelope                   avgt   20  206741.486 ±  3162.731  ns/op
BenchmarkSTContains.stContainsOuterPointInEnvelopeDeserialized       avgt   20   66869.945 ±  1507.815  ns/op
BenchmarkSTContains.stContainsOuterPointInEnvelopeSimpleGeometry     avgt   20     919.686 ±    28.572  ns/op
BenchmarkSTContains.stContainsOuterPointNotInEnvelope                avgt   20     109.417 ±     2.514  ns/op
BenchmarkSTContains.stContainsOuterPointNotInEnvelopeDeserialized    avgt   20      29.153 ±     0.734  ns/op
BenchmarkSTContains.stContainsOuterPointNotInEnvelopeSimpleGeometry  avgt   20     109.529 ±     2.475  ns/op
```