@mbasmanova 2018-01-25
@sopel,

>> It seems (from results) that optimized stContains is slightly slower in general case, but helps greatly in extreme cases.

For shapes on the left which lie within the bounding box of the shape on the right optimized ST_Contains is just under 2% slower. For shapes that don't lie within the bounding box the new version can be thousands of times faster. I think of both cases as common scenarios. Also, I'm seeing many queries where the latter case happens a lot more often than the former. Consider a spatial join `SELECT ... FROM a, b WHERE ST_Contains(a.x, b.y)` where ST_Contains is applied to all combinations of records from 'a' and 'b'. In this query only tiny subset of invocations will require checking beyond bounding boxes. A range query filtering a dataset using fixed polygon would be similar.

I looked at a concrete example where table 'a' had 800K rows and table 'b' had 34. ST_Contains is invoked 27M times. In 26M of these, bounding boxes don't match. In 500K cases bounding boxes do match. The old implementation would spend about 1000 hours (non-distributed), while the new one will spend only 28 hours.

```
  New: 204ms x 500K + 0.1ms x 26M ~= 104K seconds 
  Old: 200ms x 500K + 134ms x 26M ~= 1K hours
```