@mbasmanova 2018-01-25
@nezihyigitbasi , @sopel39 , thanks for reviews. I believe I addressed all comments.